% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{binomialTestMatrix}
\alias{binomialTestMatrix}
\title{Generate matrix of pvalues indicating if a cell is enriched in neighbors of a given cell type}
\usage{
binomialTestMatrix(cells, neighDist = 50, ncores = 1, verbose = TRUE)
}
\arguments{
\item{cells}{sf object, with celltypes features and point geometries}

\item{neighDist}{distance to define neighbors (default = 50)}

\item{ncores}{number of cores for parallelization (default 1)}

\item{verbose}{Boolean for verbosity (default TRUE)}
}
\value{
matrix where rows are cells, columns are cell types and values are p-values whether or not a cell is enriched in neighbors of a given cell type based on a binomial test.
}
\description{
pvalues are based on a binomial test and neighbors are defined within a given distance from a cell
}
\examples{
\dontrun{
data(sim)
cells <- toSF(pos = sim[,c("x", "y")], celltypes = sim$celltypes)
shuffleList <- makeShuffledCells(cells, scales = c(150, 250, 500, 750, 1000), ncores = 2)
binomMat <- binomialTestMatrix(cells, neighDist = 50, ncores = 2)
}

}
