% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_outputs.R
\name{calculateAUC}
\alias{calculateAUC}
\title{Calculate AUC for each cell type pair from a dat file}
\usage{
calculateAUC(datList, sharedPairs = TRUE)
}
\arguments{
\item{datList}{list of \code{findTrends()} data.frames; the information about the
scale, Z-score, reference and the neighbor cell. The input data.frame should
be the results list from \code{findTrends()} that has been melted into a
data.frame using \code{meltResultsList()}.}

\item{sharedPairs}{boolean; if true, will remove the AUC values from the
cell-type pairs that are not available in all samples (default: TRUE).}
}
\value{
data.frame; the AUC values for the cell-type pairs in the samples
}
\description{
Calculate the AUC values for each cell type pair from the dat variable of a
sample.
}
