% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateCelltypeProportions}
\alias{calculateCelltypeProportions}
\title{Calculate proportions of cell types inside the neighborhood}
\usage{
calculateCelltypeProportions(cells, ref, neighDist)
}
\arguments{
\item{cells}{sf data.frame; as produced by crawdad::toSF function: cells with
cell types annotated in the celltypes column and point positions in the
geometry column}

\item{ref}{character; reference cell type to define neighborhood around}

\item{neighDist}{numeric vector; distances used to define the neighborhoods}
}
\value{
named vector; the proportions
}
\description{
For a reference cell type and neighborhood distance, calculate the
proportion of each neighbor cell type inside the neighborhood of the reference
cell type at that distance.
}
