% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correctZBonferroni}
\alias{correctZBonferroni}
\title{Perform Bonferroni Multiple-test Correction}
\usage{
correctZBonferroni(dat, pSigThresh = 0.05)
}
\arguments{
\item{dat}{data.frame; the melted findTrends result.}

\item{pSigThresh}{numeric; the pvalue to be converted. Defaults to 0.05.}
}
\value{
numeric; the Bonferroni corrected zscore.
}
\description{
Calculate the number of tests performed by multiplying the number of cell
types used in the analysis. Then, divide the 0.05 pvalue by the number of
tests. Finally, convert the new pvalue to a zscore.
}
