% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparing_trends.R
\name{diffTrendTesting}
\alias{diffTrendTesting}
\title{Test if intra sample vs inter samples trends are significantly different}
\usage{
diffTrendTesting(
  samples,
  refID,
  neighID,
  zScores,
  heatmap = TRUE,
  distplot = TRUE
)
}
\arguments{
\item{samples}{list of sample set names for each patient ex: list(c("PKHL", "XXCD"),c("KSFB", "NGPL"),c("PBVN", "FSLD"))}

\item{refID}{name of reference cell type for trend of interest}

\item{neighID}{name of neighbor cell type for trend of interest}

\item{zScores}{table of zScores for each cell type combo for each sample (rows) vs each scale tested (columns)}

\item{heatmap}{return heatmap of distances between sample trends (boolean, default: TRUE)}

\item{distplot}{return plot of distirbutions of intra and inter trend distances (boolean, default: TRUE)}
}
\value{
pvalues for each patient
}
\description{
Computes summed distance between z scores at each scale for two given trends.
Gets summed distances for combinations of intra patient trends, and inter patient trends.
T-test between intra and inter distances to test for significant difference.
Also can return heatmap matrix of distances between each dataset and distributions.
The idea is that if inter trends are similar, then their distances will be small like the intra trends.
But if the inter trends are different, their distances should be larger than the intra trends.
Likewise, if the intra trends are highly variable, then might be expected to also have large differences between inter as well.
}
\examples{
\dontrun{
diffTrendTesting(list(c("PKHL", "XXCD"),c("KSFB", "NGPL"),c("PBVN", "FSLD")),
                 refID = "Ki67 proliferating",
                 neighID = "CD4 Memory T cells",
                 zScores = combined_zscores)
}

}
