% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_inputs.R
\name{filterCells}
\alias{filterCells}
\title{remove celltypes that have few cells in the datset}
\usage{
filterCells(cells, thresh = 0, verbose = TRUE)
}
\arguments{
\item{cells}{sp::SpatialPointsDataFrame object, with celltypes features column}

\item{thresh}{threshold to remove cell types that contribute less than this fraction to the dataset (default: 0.0)}

\item{verbose}{Boolean; verbosity (default TRUE)}
}
\description{
remove celltypes that have few cells in the datset
}
