% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_outputs.R
\name{filterSepTrends}
\alias{filterSepTrends}
\title{filter for significant cell type association trends that are separated}
\usage{
filterSepTrends(results, alpha = 0.05)
}
\arguments{
\item{results}{list output from \code{findTrends()}}

\item{alpha}{significance threshold}
}
\description{
filter the results list from \code{findTrends()} for neighbor cell types that are significantly separated with each reference cell type.
}
