% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{findTrends}
\alias{findTrends}
\title{Compute trends of cell type colocalization for each cell type combination across specified resolutions}
\usage{
findTrends(
  pos,
  celltypes,
  resolutions = c(50, 100, 200, 300),
  dist = 50,
  sub.dist = 100,
  sub.type = c("pairwise", "near", "away"),
  sub.thresh = 0.05,
  perms = 1,
  seed = 0,
  ncores = 1,
  loadShuffleFile = NA,
  saveShuffleFilePath = NA,
  loadSubsetFile = NA,
  saveSubsetFile = NA,
  plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{pos}{data frame; x and y coordinates of each cell}

\item{celltypes}{character vector; the cell type of each cell provided in pos}

\item{resolutions}{numeric vector; the different resolutions to shuffle at and subsequently compute significance at}

\item{dist}{numeric; distance to define neighbor cells with respect to each reference cell (default = 50)}

\item{sub.dist}{numeric; distance to define subsets relative to a neighbor cell type (default = 100)}

\item{sub.type}{character; subset type, either "pairwise", to not use subsets, or subsets of ref cells "near" (ie localized) a neighbor cell type, or "away" (ie separated) from a neighbor cell type}

\item{sub.thresh}{numeric; significance threshold for the binomial test (default = 0.05)}

\item{perms}{numeric; number of permutations to shuffle for each resolution (default = 1)}

\item{seed}{numeric; set seed for shuffling (if more than 1 permutation, then seed equals permutation number)}

\item{ncores}{numeric; number of cores for parallelization (default 1)}

\item{loadShuffleFile}{character; path to a preshuffled randomcellslist rds object (default NA)}

\item{saveShuffleFilePath}{character; can save the shuffled cell labels to speed things up later (default NA)}

\item{loadSubsetFile}{character; path to a premade subset list rds object (default NA)}

\item{saveSubsetFile}{character; can save the subset cell list to speed things up later, or use for subsequent plotting (default NA)}

\item{plot}{Boolean; return plots (default TRUE)}

\item{verbose}{Boolean; verbosity (default TRUE)}
}
\value{
A list that contains a dataframe for each reference cell type, where the dataframe contains the significance values for each neighbor cell type at each resolution
}
\description{
Trends are based on significant differences in cell type proportions between the real and randomly shuffled datasets.
Cell type proportions are with respect to the different cell types that are neighboring the cells of a given reference cell type within a certain defined distance.
This is done at difference resolutions, where a resolution is whether the cell type labels are shuffled locally or globally.
Trends are essentially built from significance values. The significance test basically asks if two cell types are localized or separated by assessing if the proportion of the neighboring cell type is significantly greater, or less than, random chance.
}
