% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{getSubsets}
\alias{getSubsets}
\title{Find subsets of cells}
\usage{
getSubsets(
  cells,
  sub.dist = 100,
  sub.type = c("near", "away"),
  sub.thresh = 0.05,
  ncores = 1,
  verbose = TRUE
)
}
\arguments{
\item{cells}{sp::SpatialPointsDataFrame object, with celltypes features and point geometries}

\item{sub.dist}{distance to define subsets relative to a neighbor cell type (default = 50)}

\item{sub.type}{subset type, either ref cells "near" (ie localized) a neighbor cell type, or "away" (ie separated) from a neighbor cell type.}

\item{sub.thresh}{significance threshold for the binomial test (default = 0.05)}

\item{ncores}{number of cores for parallelization (default 1)}

\item{verbose}{Boolean for verbosity (default TRUE)}
}
\value{
list where each entry is a subset and the values are the cell ids determined to be in the subset
}
\description{
find the subset cells of a reference cell type defined by being either significantly "near" or "away" with respect to a given neighbor cell type.
Significantly "near" or "away" defined by whether the proportion of adjacent cells within a given distance are either enriched or depleted in a given neighbor cell type compared to the global proportion of the given neighbor cell type.
Note that for near, or localized, just test if significantly enriched.
For "away", do the same, but then take the cells that were not significant. However, would recommend setting the p-value threshold to be very liberal, like 0.5, that way, only the cells that couldn't even pass a p-val cutoff of 0.5 would be selected for, and these would be expected to be very much depleted or separated from the neighbor cell type.
}
