% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{makeShuffledCells}
\alias{makeShuffledCells}
\title{Shuffle the celltype labels at different resolutions}
\usage{
makeShuffledCells(
  cells,
  resolutions,
  perms = 1,
  ncores = 1,
  seed = 0,
  verbose = TRUE
)
}
\arguments{
\item{cells}{sp::SpatialPointsDataFrame object; celltypes features and point geometries}

\item{resolutions}{numeric vector; the different resolutions to shuffle at and subsequently compute significance at}

\item{perms}{numeric; number of permutations to shuffle for each resolution (default 1)}

\item{ncores}{numeric; number of cores for parallelization (default 1)}

\item{seed}{numeric; set seed for shuffling (if more than 1 permutation, then seed equals permutation number)}

\item{verbose}{Boolean; verbosity (default TRUE)}
}
\value{
list;
}
\description{
For each resolution, and for i number of permutations, shuffle the celltype labels.
Returns a list of lists, where each list is a factor of the shuffled celltype labels.
Outer list is for each resolution of shuffling, and each inner list is for a given permutation with a given seed.
}
