% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkhl}
\alias{pkhl}
\title{Preprocessed CODEX metadata of a human spleen.}
\format{
data.frame object where rows are cells,
\code{x} and \code{y} are x-y coordinates,
\code{celltypes} are cell type annotations
}
\source{
\url{https://portal.hubmapconsortium.org/browse/dataset/1393ff7729eff13cd2f4d5d58af96fdb}
}
\usage{
data(pkhl)
}
\description{
Preprocessed CODEX metadata of a human spleen.
}
\keyword{datasets}
