% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTrends}
\alias{plotTrends}
\title{Plot trends}
\usage{
plotTrends(
  results,
  idcol = "id",
  figPath = "results.pdf",
  width = 8,
  height = 8,
  legend = TRUE,
  ...
)
}
\arguments{
\item{idcol}{if results are a data.frame, this is the column that contains the additional feature to plot multiple trend lines with
by default it is 'id', which would be the name given to the column describing each list after melting together into a single dataframe
using `meltResultsList()`}

\item{...}{additional plotting parameters for base R plotting. Fed into "lines()" in script}
}
\description{
Plot calculated trends
}
