% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTrendsOverlay}
\alias{plotTrendsOverlay}
\title{This one overlays each neighbor trend wrt the same reference cell type on the plot}
\usage{
plotTrendsOverlay(
  results,
  figPath = "results.pdf",
  width = 4,
  height = 10,
  legend = TRUE,
  ...
)
}
\arguments{
\item{...}{additional plotting parameters for base R plotting. Fed into "lines()" in script}
}
\description{
This one overlays each neighbor trend wrt the same reference cell type on the plot
}
