% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{selectLabels}
\alias{selectLabels}
\title{Update cell type labels to only include specific cell types or subtypes}
\usage{
selectLabels(df, com, cellIDs = NA, subset_list = NA, subsetIDs = NA)
}
\arguments{
\item{df}{dataframe or sp::SpatialPointsDataFrame object of cells}

\item{com}{original factor or vector of cell type labels for cells in \code{df}}

\item{cellIDs}{vector of cell type labels to include in output (default: NA)}

\item{subset_list}{list of subsets from \code{selectSubsets()} (default; NA)}

\item{subsetIDs}{vector of susbet cell type labels to include in output (names of vectors in \code{subset_list}) (default: NA)}
}
\value{
factor of specific cell type labels for visualizing with \code{vizAllClusters()} in the parameter: \code{clusters}
}
\description{
make a factor of selected cell type labels for visualizing specific cells with \code{vizAllClusters()}.
Could append new entries into the \code{subset_list} to select and label other custom subsets of cells
}
