% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{selectSubsets}
\alias{selectSubsets}
\title{find subsets of cells}
\usage{
selectSubsets(
  binomMatrix,
  celltypes,
  sub.type = c("near", "away"),
  sub.thresh = 0.05,
  ncores = 1,
  verbose = TRUE
)
}
\arguments{
\item{binomMatrix}{matrix where rows are cells, columns are cell types and values are p-values whether or not a cell is enriched in neighbors of a given cell type based on a binomial test. Output from \code{binomialTestMatrix()}}

\item{celltypes}{named vector or factor of cell type labels of each cell in \code{binomMatrix} and in the same order.}

\item{sub.type}{subset type, either ref cells "near" (ie localized) a neighbor cell type, or "away" (ie separated) from a neighbor cell type.}

\item{sub.thresh}{significance threshold for the binomial test (default = 0.05)}

\item{ncores}{number of cores for parallelization (default 1)}

\item{verbose}{Boolean for verbosity (default TRUE)}
}
\value{
list where each entry is a subset and the values are the cell ids determined to be in the subset
}
\description{
assign cells to subsets based on whether they are enriched in a given neighbor cell type based on the pvalues in the input \code{binomMatrix}
}
