% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seq}
\alias{seq}
\title{Preprocessed subset of seqFISH metadata of a mouse embryo}
\format{
data.frame object where rows are cells,
\code{x} and \code{y} are x-y coordinates,
\code{celltypes} are cell type annotations
}
\source{
\url{https://www.nature.com/articles/s41587-021-01006-2}
}
\usage{
data(seq)
}
\description{
Preprocessed subset of seqFISH metadata of a mouse embryo
}
\keyword{datasets}
