% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_inputs.R
\name{seuratToSF}
\alias{seuratToSF}
\title{Convert Seurat object to an sf object}
\usage{
seuratToSF(obj, coms, posIDs = c("x", "y"), verbose = TRUE)
}
\arguments{
\item{obj}{the seurat object}

\item{coms}{name of \code{meta.dat}a column that contains cell annotations}

\item{posIDs}{columns in the \code{meta.data} that contain the spatial coordinates (default: c("x", "y"))}

\item{verbose}{Boolean; verbosity (default TRUE)}
}
\description{
Assumes that cell spatial coordinates and cell type annotations/communities are columns in the \code{meta.data} slot
}
