% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{Simulation}
\format{
data.frame object where rows are cells,
\code{x} and \code{y} are x-y coordinates,
\code{celltypes} are cell type annotations
}
\usage{
data(sim)
}
\description{
Simulation
}
\keyword{datasets}
