% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_circles}
\alias{simulate_circles}
\title{Create cell type circle patterns in the background of cells}
\usage{
simulate_circles(pos, locs, radii, cts, probs)
}
\arguments{
\item{pos}{dataframe; the background where the patter will be created. It should have columns x, y, and type (typically from simulate_background)}

\item{locs}{list of 2-d numeric vectors; each vector is the x and y coordinates of a circle center (0-1)}

\item{radii}{list of 2-d lists, where each inner list has variables "inner" and "outer" that refer to the
radius of the outer ring and the inner cores of each circle. An inner list for each circle in locs}

\item{cts}{same format as radii, but "inner" and "outer" are vectors of cell types present in core and ring}

\item{probs}{same format as radii, but "inner" and "outer" are vectors of cell types proportions}
}
\description{
takes a dataframe of x, y, and cell type (from simulate_background)
and selects cells to create new cell type patterns of circles.
The circles can actually be an outer ring and an inner core.
}
