% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slide}
\alias{slide}
\title{Preprocessed subset of Slide-seq metadata of a mouse cerebellum.
slideseq.puck190926_08 used in RCTD (https://raw.githack.com/dmcable/spacexr/master/AnalysisPaper/MainFigures/figure4.html)}
\format{
data.frame object where rows are cells,
\code{x} and \code{y} are x-y coordinates,
\code{celltypes} are cell type annotations
}
\source{
\url{https://www.dropbox.com/sh/6klgiz1f9rimzqb/AADX_ziLibO7XY5S4rwPLdQoa/Figure4?dl=0&subfolder_nav_tracking=1}
}
\usage{
data(slide)
}
\description{
Preprocessed subset of Slide-seq metadata of a mouse cerebellum.
slideseq.puck190926_08 used in RCTD (https://raw.githack.com/dmcable/spacexr/master/AnalysisPaper/MainFigures/figure4.html)
}
\keyword{datasets}
