% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_inputs.R
\name{spToDF}
\alias{spToDF}
\title{Convert an sf object to a dataframe with x and y coordinates and cell type labels}
\usage{
spToDF(cells)
}
\arguments{
\item{cells}{sf object, with celltypes features and point geometries}
}
\value{
datafame with columns: x, y, and celltype
}
\description{
Convert an sf object to a dataframe with x and y coordinates and cell type labels
}
