% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_inputs.R
\name{toSF}
\alias{toSF}
\title{Convert positions and cell type annotations to sf object}
\usage{
toSF(pos, cellTypes, verbose = TRUE)
}
\arguments{
\item{pos}{data frame; x and y coordinates of each cell}

\item{cellTypes}{character vector; the cell type of each cell provided in pos}

\item{verbose}{Boolean; verbosity (default TRUE)}
}
\description{
Convert positions and cell type annotations to sf object
}
\examples{
\dontrun{
data(sim)
cells <- toSF(df)
}

}
