% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_inputs.R
\name{toSP}
\alias{toSP}
\title{Go from positions and celltype annotations to spatialpointsdataframe object}
\usage{
toSP(pos, celltypes, verbose = TRUE)
}
\arguments{
\item{pos}{data frame; x and y coordinates of each cell}

\item{celltypes}{character vector; the cell type of each cell provided in pos}

\item{verbose}{Boolean; verbosity (default TRUE)}
}
\description{
Go from positions and celltype annotations to spatialpointsdataframe object
}
\examples{
\dontrun{
data(sim)
cells <- toSP(pos = sim[,c("x", "y")], celltypes = slide$type)
}

}
