% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizCelltypeProportions}
\alias{vizCelltypeProportions}
\title{Plot Cell-type Proportions}
\usage{
vizCelltypeProportions(cells, neighDist)
}
\arguments{
\item{cells}{sf data.frame; as produced by crawdad::toSF function: cells with
cell types annotated in the celltypes column and point positions in the
geometry column}

\item{neighDist}{numeric; distance used to define the neighborhood}
}
\value{
ggplot2 plot; the a histogram of the proportions
}
\description{
For a chosen a neighborhood distance, plot a histogram of
the proportion of each neighbor cell type inside the neighborhood of the
reference cell type for all reference cell types.
}
