% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizClusters}
\alias{vizClusters}
\title{Visualize clusters in the same plot}
\usage{
vizClusters(
  cells,
  ofInterest = NULL,
  pointSize = 1,
  alpha = 0.5,
  ref = NULL,
  dist = NULL,
  lineWidth = 0.1
)
}
\arguments{
\item{cells}{sf object; spatial (x and y) coordinates and celltypes column}

\item{ofInterest}{character vector; a vector of specific clusters to visualize}

\item{pointSize}{numeric; size of points}

\item{alpha}{numeric; transparency of points}

\item{ref}{character; reference cell type to draw the neighborhood
around. If NULL, it will not create the neighborhood (default: NULL)}

\item{dist}{numeric; distance to define neighbor cells with respect to each
reference cell. If NULL, it will not create the neighborhood (default: NULL)}

\item{lineWidth}{numeric; width of neighborhood line}
}
\value{
plot
}
\description{
Uses the cells sf object to visualize the clusters together in
the same plot.
}
\examples{
\dontrun{
data(slide)
cells <- crawdad::toSF(pos = slide[,c("x", "y")], celltypes = slide$celltypes)
vizClusters(cells)
}

}
