% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizColocDotplot}
\alias{vizColocDotplot}
\title{Plot Co-localization Dotplot}
\usage{
vizColocDotplot(
  dat,
  zSigThresh = 1.96,
  pSigThresh = NULL,
  zScoreLimit = NULL,
  reorder = FALSE,
  symmetrical = FALSE,
  onlySignificant = FALSE,
  colors = c("blue", "white", "red"),
  dotSizes = c(6, 31)
)
}
\arguments{
\item{dat}{\code{findTrends()} data.frame; the information about the scale,
Z-score, reference and the neighbor cell. The input data.frame should be the
results list from \code{findTrends()} that has been melted into a data.frame
using \code{meltResultsList()}.}

\item{zSigThresh}{numeric; the Z score significance threshold (default: 1.96).}

\item{pSigThresh}{numeric; the two-sided P value significance threshold. It
can be used in place of the zSigThresh parameter. If no value is provided,
the zSigThresh will be used.}

\item{zScoreLimit}{numeric; limit the Z-score to look better in the graph
scale gradient. Z-score values above zScoreLimit will be represented as
zScoreLimit, scores below -zScoreLimit will be represented as -zScoreLimit
(default: NULL).}

\item{reorder}{boolean; if TRUE, reorder the cell types by clustering on the
z-score. If false, orders in alphabetical order (default: FALSE).}

\item{symmetrical}{boolean; highligh relationships that are symmetrical between cell
type pairs (default: TRUE).}

\item{onlySignificant}{boolean; plot only cell types with significant
relationships (default: FALSE).}

\item{colors}{character vector; colors for the gradient heatmap (low, mid, high)
(default: c("blue", "white", "red")).}

\item{dotSizes}{numeric vector; minimum and maximum size of the dot
(default: c(6,31)).}
}
\description{
This function takes the \code{findTrends()} melted data frame and
plots the scale and the Z-score in which the trend first crossed the
significance line (Z = 1.96). The Z-score was capped between -3 and 3. Since
the co-localization at smaller scales are more important than those at
greater ones, we plotted the inverse of the scale so smaller ones would
correspond to larger dots.
}
\examples{
\dontrun{
data(sim)
cells <- toSF(pos = sim[,c("x", "y")], celltypes = sim$celltypes)
shuffle.list <- makeShuffledCells(cells, scales = c(150, 250, 500, 750, 1000), ncores = 2)
results <- findTrends(cells, dist = 100, shuffle.list = shuffle.list, ncores = 2)
dat <- meltResultsList(results)
vizColocDotplot(dat)
}

}
