% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizEachCluster}
\alias{vizEachCluster}
\title{Visualize each cluster separately}
\usage{
vizEachCluster(
  cells,
  coms,
  axisAdj = 1,
  s = 0.5,
  a = 1,
  nacol = transparentCol(color = "gray", percent = 50),
  clustcol = "red"
)
}
\arguments{
\item{cells}{either a data.frame or sp::SpatialPointsDataFrame object with cell spatial coordinates}

\item{coms}{a factor of cell type labels for the cells}

\item{axisAdj}{how much to increase axis ranges. If tissue, 100 okay, if embedding, 1 ok (default: 100)}

\item{a}{alpha of points (default: 1; no transparency)}

\item{nacol}{color of the NA values for cells of "other" cluster (default: (transparentCol(color = "gray", percent = 50)))}

\item{clustcol}{color of the cells of the given cluster (default: red)}

\item{size}{size of points (default: 0.01)}
}
\value{
plot where each panel is a different celltype
}
\description{
Returns a gridExtra of grobs.
A single plot where each panel is a different cluster highlighted on the tissue
}
\examples{
\dontrun{
data(slide)
vizEachCluster(slide, coms = slide$celltypes)
}

}
