% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizGrids}
\alias{vizGrids}
\title{Visualize grids and clusters}
\usage{
vizGrids(
  cells,
  scale,
  permutation = 1,
  totalPermutations = NULL,
  square = TRUE,
  ofInterest = NULL,
  pointSize = 1,
  alpha = 0.5
)
}
\arguments{
\item{cells}{sf object; spatial (x and y) coordinates and celltypes column}

\item{scale}{numeric; size of the scale to plot}

\item{permutation}{numeric; the number of the permutation of interest.}

\item{totalPermutations}{numeric; the total number of permutations used to
shuffle the data.}

\item{square}{boolean; if true, create a squared grid, if false, make
hexagonal grid (default TRUE)}

\item{ofInterest}{character vector; a vector of specific clusters to visualize}

\item{pointSize}{numeric; size of points}

\item{alpha}{numeric; transparency of points}
}
\value{
plot
}
\description{
Uses the cells sf object and size of grid to visualize the grids
used to create the null background.
}
\examples{
\dontrun{
data(slide)
cells <- crawdad::toSF(pos = slide[,c("x", "y")], celltypes = slide$celltypes)
vizAllClusters(cells)
}

}
