% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizPCASamples}
\alias{vizPCASamples}
\title{Visualize the samples using their AUC and PCA}
\usage{
vizPCASamples(aucSamples)
}
\arguments{
\item{aucSamples}{data.frame; the AUC values for each cell-type pair in each
sample as calculated by \code{calculateAUC()}}
}
\value{
gglot2 plot; the PCA visualization of the samples
}
\description{
Uses the AUC values calculated from each cell-type pair in each sample to
represent the samples in the PCA reduced dimension space.
}
