% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizTrends.heatmap}
\alias{vizTrends.heatmap}
\title{Plot trends as heatmaps with ggplot2}
\usage{
vizTrends.heatmap(
  dat,
  sig.thresh = 1.96,
  z_limit = 20,
  palette_ = (grDevices::colorRampPalette(c("blue", "white", "red")))(n = 209),
  title = NULL,
  withPerms = FALSE,
  annotation = FALSE,
  ncols = 4
)
}
\arguments{
\item{dat}{\code{findTrends()} results list, or data.frame; the information about the scale, Z-score, reference and the neighbor cell.}

\item{sig.thresh}{threshold for significance, ie Z score significance threshold (default: 1.96).}

\item{z_limit}{Z score limits (default +/- 20)}

\item{palette_}{color gradient for heatmap (default: grDevices::colorRampPalette(c("blue", "white", "red"))(n = 209))}

\item{title}{plot title (default: NULL)}

\item{withPerms}{if the results list is a list of lists using \code{returnMeans = FALSE} in \code{findTrends()}, then column order is different and this flag is needed (default: FALSE)}

\item{annotation}{boolean to show the Z score values in the squares of the heatmap (default: FALSE)}

\item{ncols}{specify number of columns in facet wrap (default: 4)}
}
\description{
The input data.frame should be the results list from \code{findTrends()} that has been melted into a data.frame using \code{meltResultsList()}.
}
\examples{
\dontrun{
data(sim)
cells <- toSP(pos = sim[,c("x", "y")], celltypes = slide$type)
shuffle.list <- makeShuffledCells(cells, scales = c(150, 250, 500, 750, 1000, 1500, 2000), ncores = 2)
results <- findTrends(cells, dist = 100, shuffle.list = shuffle.list, ncores = 2)
vizTrends.heatmap(dat = results)
}

}
