% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizTrends}
\alias{vizTrends}
\title{Plot trends with ggplot2}
\usage{
vizTrends(
  dat,
  id = "id",
  yaxis = "Z",
  sig.thresh = 2,
  nc = length(unique(dat[[id]])),
  colors = rainbow(nc),
  title = NULL,
  facet = TRUE,
  lines = TRUE,
  points = TRUE
)
}
\arguments{
\item{dat}{data.frame; the information about the resolution, Z-score, reference and the neighbor cell.}

\item{id}{column name that contains an additional feature to color trend lines (default: "id")}

\item{yaxis}{column that has significance value across resolutions (default: "Z")}

\item{sig.thresh}{threshold for significance, ie Z score significance threshold (default: 2).}

\item{nc}{number of colors to use for labeling features in id column}

\item{colors}{color assignment for each of the features in id column}

\item{title}{plot title (default: NULL)}

\item{facet}{boolean to facet wrap on reference and neighbor cell types. (default: TRUE)}

\item{lines}{boolean to plot lines (default: TRUE)}

\item{points}{boolean to plot points (default: TRUE)}
}
\description{
The input data.frame should be the results list from \code{findTrends()} that has been melted into a data.frame using \code{meltResultsList()}.
}
