% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizVarianceSamples}
\alias{vizVarianceSamples}
\title{Visualize the variance of each cell-type relationship in the samples}
\usage{
vizVarianceSamples(aucSamples)
}
\arguments{
\item{aucSamples}{data.frame; the AUC values for each cell-type pair in each
sample as calculated by \code{calculateAUC()}}
}
\value{
gglot2 plot; the dot plot visualization of the variance in the
samples
}
\description{
Visualize the variance of the AUC values for each cell-type pair in all
samples.
}
