% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{binomialTestMatrix}
\alias{binomialTestMatrix}
\title{Generate matrix of pvalues indicating if a cell is enriched in neighbors of a given cell type}
\usage{
binomialTestMatrix(cells, neigh.dist = 100, ncores = 1, verbose = TRUE)
}
\arguments{
\item{cells}{sp::SpatialPointsDataFrame object, with celltypes features and point geometries}

\item{neigh.dist}{distance to define neighbors (default = 100)}

\item{ncores}{number of cores for parallelization (default 1)}

\item{verbose}{Boolean for verbosity (default TRUE)}
}
\value{
matrix where rows are cells, columns are cell types and values are p-values whether or not a cell is enriched in neighbors of a given cell type based on a binomial test.
}
\description{
pvalues are based on a binomial test and neighbors are defined within a given distance from a cell
}
\examples{
\dontrun{
data(sim)
cells <- toSP(pos = sim[,c("x", "y")], celltypes = slide$type)
shuffle.list <- makeShuffledCells(cells, resolutions = c(150, 250, 500, 750, 1000, 1500, 2000), ncores = 2)
binomMat <- binomialTestMatrix(cells, neigh.dist = 100, ncores = 2)
}

}
