% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{evaluateSignificance}
\alias{evaluateSignificance}
\title{Compute significant different between real and randomly shuffled cell neighbor proportions}
\usage{
evaluateSignificance(
  cells,
  randomcellslist,
  trueNeighCells,
  cellBuffer,
  ncores = 1,
  removeDups = TRUE,
  returnMeans = TRUE
)
}
\arguments{
\item{cells}{sf object of all the cells}

\item{randomcellslist}{list of lists of randomly shuffled cell type labels produced from \code{makeShuffledCells}}

\item{trueNeighCells}{Simple feature collection of real cells for a given reference cell type, with geometries of a given neighDist (from sf::st_buffer)}

\item{cellBuffer}{Simple feature collection of the neighbor cells that are within "neighDist" of the ref cells (from sf::intersection)}

\item{ncores}{number of cores for parallelization (default 1)}

\item{removeDups}{remove duplicate neighbor cells to prevent them from being counted multiple times and inflate the Z scores (default: TRUE)}

\item{returnMeans}{if multiple permutations, return the mean Z score across the permutations in each scale with respect to each neighbor cell type (default: TRUE)}
}
\description{
for a given reference cell type, computes the Z scores for being colocalized or separated from each query cell type.
If \code{returnMeans = TRUE}, then the result will be a data.frame where each row is a scale, each column is a query cell type, and each value is the Z score.
If \code{returnMeans = FALSE}, then the result will be a list of data.frames, where each data.frame is a scale,
the rows are permutations, the columns are query cell types, and each value is a Z score.
}
