% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_outputs.R
\name{filterCoTrends}
\alias{filterCoTrends}
\title{filter for significant cell type association trends that are co-localized}
\usage{
filterCoTrends(results, alpha = 0.05)
}
\arguments{
\item{results}{list output from \code{findTrends()}}

\item{alpha}{significance threshold}
}
\description{
filter the results list from \code{findTrends()} for neighbor cell types that are significantly co-localized with each reference cell type.
}
\examples{
\dontrun{
data(sim)
cells <- toSF(pos = sim[,c("x", "y")], celltypes = sim$celltypes)
shuffle.list <- makeShuffledCells(cells, scales = c(150, 250, 500, 750, 1000), ncores = 2)
results <- findTrends(cells, dist = 100, shuffle.list = shuffle.list, ncores = 2)
filterCoTrends(results = results, alpha = 0.05)
}

}
