% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{findTrends}
\alias{findTrends}
\title{Compute trends of cell type colocalization for each cell type combination across specified resolutions}
\usage{
findTrends(
  cells,
  dist = 50,
  sub.dist = 100,
  sub.type = c("pairwise", "near", "away"),
  sub.thresh = 0.05,
  perms = 1,
  seed = 0,
  ncores = 1,
  shuffle.list,
  subset.list = NA,
  plot = FALSE,
  verbose = TRUE,
  removeDups = TRUE
)
}
\arguments{
\item{cells}{sp::SpatialPointsDataFrame object, with celltypes features and point geometries}

\item{dist}{numeric distance to define neighbor cells with respect to each reference cell (default: 50)}

\item{sub.dist}{distance to define subsets relative to a neighbor cell type (default = 100)}

\item{sub.type}{subset type, either "pairwise", to not use subsets, or susbets of ref cells "near" (ie localized) a neighbor cell type, or "away" (ie separated) from a neighbor cell type.}

\item{sub.thresh}{significance threshold for the binomial test (default = 0.05)}

\item{perms}{number of permutations to shuffle for each resolution (default = 1)}

\item{seed}{set seed for shuffling (if more than 1 permutation, then seed equals permutation number)}

\item{ncores}{number of cores for parallelization (default 1)}

\item{shuffle.list}{a list of cell type labels shuffled at different resolutions (output from \code{makeShuffledCells()})}

\item{subset.list}{a subset list (output from \code{selectSubsets()}). Required if computing trends for subsets (default NA)}

\item{plot}{Boolean to return plots (default TRUE)}

\item{verbose}{Boolean for verbosity (default TRUE)}
}
\value{
A list that contains a dataframe for each reference cell type, where the dataframe contains the significance values for each neighbor cell type at each resolution
}
\description{
Trends are based on significant differences in cell type proportions between the real and randomly shuffled datasets.
Cell type proportions are with respect to the different cell types that are neighboring the cells of a given reference cell type within a certain defined distance.
This is done at difference resolutions, where a resolution is whether the cell type labels are shuffled locally or globally.
Trends are essentially built from significance values. The significance test basically asks if two cell types are localized or separated by assessing if the proportion of the neighboring cell type is significantly greater, or less than, random chance.
}
