% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_trends.R
\name{makeShuffledCells}
\alias{makeShuffledCells}
\title{Shuffle the celltype labels at different shuffling scales}
\usage{
makeShuffledCells(
  cells,
  scales = c(50, 100, 200, 300, 400, 500),
  perms = 5,
  ncores = 1,
  seed = 0,
  square = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cells}{sf object, with celltypes features and point geometries}

\item{scales}{numeric vector of the different scales to shuffle at and
subsequently compute significance at (default: c(50, 100, 200, 300, 400, 500))}

\item{perms}{number of permutations to shuffle for each scale (default 5)}

\item{ncores}{number of cores for parallelization (default 1)}

\item{seed}{set seed for shuffling (if more than 1 permutation, each
shuffling permutation has a seed equal to the permutation number)}

\item{square}{if false, make hexagonal grid (default TRUE)}

\item{verbose}{Boolean for verbosity (default TRUE)}
}
\description{
for each scale, and for i number of permutations, shuffle the celltype labels.
Returns a list of lists, where each list is a factor of the shuffled celltype labels.
Outer list is for each scale of shuffling, and each inner list is for a given
permutation with a given seed.
}
\examples{
 
\dontrun{
data(sim)
makeShuffledCells(sim, scales = c(50, 100, 200, 300, 400, 500), ncores = 2)
}

}
