% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTrendsOverlay}
\alias{plotTrendsOverlay}
\title{This one overlays each neighbor trend wrt the same reference cell type on the plot}
\usage{
plotTrendsOverlay(results, figpath = NULL, width = 4, height = 4, ...)
}
\arguments{
\item{results}{data.frame; the information about the resolution, Z-score, reference and the neighbor cell. It can be the result directly obtained by the melted \code{findTrends} output created by the \code{meltResultsList()} function.}

\item{figpath}{path to save plot to if desired, otherwise plot generated in R session (default: NULL)}

\item{width}{param controlling size of plot if it is to be saved to a file (default: 4)}

\item{height}{param controlling size of plot if it is to be saved to a file (default: 4)}

\item{...}{additional plotting parameters for base R plotting. Fed into "lines()" in script}
}
\value{
nothing
}
\description{
This one overlays each neighbor trend wrt the same reference cell type on the plot
}
