% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_background}
\alias{simulate_background}
\title{Simulate background}
\usage{
simulate_background(
  size = 10000,
  cts = c("A"),
  prob = c(1),
  seed = 1,
  scale = 1
)
}
\arguments{
\item{size}{numeric; number of points to generate}

\item{cts}{character vector; the different cell types in the background, vector of IDs}

\item{prob}{numeric vetor; the probability, or proportion of each cell type, summing to 1, a vector of probabilities}

\item{seed}{numeric; pseudo random seed for the function sample()}

\item{scale}{numeric; amount to expand the coordinates by}
}
\value{
data.frame; the x, y coordinates of each point, and column type indicating its cell type
}
\description{
Create a uniform background of points, where each point is a given cell type.
}
