% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{vizClusters}
\alias{vizClusters}
\title{Visualize clusters in the same plot}
\usage{
vizClusters(cells, ofInterest = NULL, pointSize = 1, alpha = 0.5)
}
\arguments{
\item{cells}{sf object; spatial (x and y) coordinates and celltypes}

\item{ofInterest}{character vector; a vector of specific clusters to visualize}

\item{pointSize}{numeric; size of points}

\item{alpha}{numeric; transparency of points}
}
\value{
plot
}
\description{
Uses the cells sf object to visualize the clusters together in
the same plot.
}
\examples{
\dontrun{
data(slide)
cells <- crawdad::toSF(pos = slide[,c("x", "y")], celltypes = slide$celltypes)
vizAllClusters(cells)
}

}
